requirements.txt
appdirs==1.4.3
Django==1.9
django-crispy-forms==1.6.0
olefile==0.44
packaging==16.8
Pillow==4.1.1
pyparsing==2.2.0
six==1.10.0
